backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
python-multipart==0.0.9
python-dotenv==1.0.1
mistralai==1.2.5
chromadb==0.5.0
pydantic==2.7.1
aiofiles==23.2.1
sentence-transformers==2.7.0
psycopg2-binary==2.9.9